fix bugs of newDay and addPrivateContest

+++ Models/problemset.php
@@ -12,10 +12,12 @@ class problemset
         $dir = "data/dataBackups/";
         $array_map = array_map('filemtime', ($files = glob($dir . "data*.txt")));
         array_multisort($array_map, SORT_ASC, $files);
-        $text = file_get_contents("data/data.txt");
-        file_put_contents("data/dataBackups/data" . Date(date(" Y.m.d h:m:s")) . ".txt", $text);
-        if (count($files) > MAXIMUM_BACKUP_FILES) {
-            unlink($files[0]);
+        if(file_exists("data/data.txt")) {
+            $text = file_get_contents("data/data.txt");
+            file_put_contents("data/dataBackups/data" . Date(date(" Y.m.d h:m:s")) . ".txt", $text);
+            if (count($files) > MAXIMUM_BACKUP_FILES) {
+                unlink($files[0]);
+            }
         }
     }
